+++ src/Services/Telephony/Call/Result/TranscriptAttachItemResult.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\Telephony\Call\Result;
+
+use Bitrix24\SDK\Core\Result\AbstractItem;
+use Bitrix24\SDK\Services\Telephony\Common\CrmEntity;
+
+/**
+ * @property-read  non-negative-int $TRANSCRIPT_ID
+ */
+class TranscriptAttachItemResult extends AbstractItem
+{
+}