+++ src/Services/Telephony/TelephonyServiceBuilder.php
@@ -28,4 +28,16 @@ class TelephonyServiceBuilder extends AbstractServiceBuilder
 
         return $this->serviceCache[__METHOD__];
     }
+    public function call(): Telephony\Call\Service\Call
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Telephony\Call\Service\Call(
+                new Telephony\Call\Service\Batch($this->batch, $this->log),
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
 }